Mention license in readme and add license file.
Previously, the fact that this project is licensed under Apache-2.0
could only be gathered from looking at the license field in
package.json. Mentioning the license in the readme helps users who are
not familiar with npm, and the additional license file should make the
license show up correctly in github's web interface aswell.